Add debug prints to sim test

+++ src/impl/random_variable.c
@@ -623,6 +623,9 @@ void rvs_reset(RandomVariables *rvs, const RandomVariablesArgs *rvs_args) {
 }
 
 void rvs_destroy(RandomVariables *rvs) {
+  if (rvs->destroy_data_func == NULL) {
+    return;
+  }
   rvs->destroy_data_func(rvs);
   free(rvs);
 }
